apps/your-nova-app/pyproject.toml
[project]
name = "your_nova_app"
version = "0.0.1"
description = ""
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "wandelbots-nova>=1.17.0,<2",
    "python-decouple~=3.8",
    "requests>=2.32.3,<3",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "mypy>=1.16.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
